requirements.txt
apache-airflow
google-cloud-storage
google-cloud-bigquery
pandas
db-dtypes
google-cloud-bigquery-storage